Restructure component into directory.
