+++ assets/js/components/notifications/EnableAutoUpdateBannerNotification/index.js
@@ -26,17 +26,17 @@ import { Fragment, useCallback, useEffect, useState } from '@wordpress/element';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
-import { getTimeInSeconds } from '../../util';
-import useQueryArg from '../../hooks/useQueryArg';
+import { CORE_SITE } from '../../../googlesitekit/datastore/site/constants';
+import { getTimeInSeconds } from '../../../util';
+import useQueryArg from '../../../hooks/useQueryArg';
 import {
 	CORE_USER,
 	PERMISSION_UPDATE_PLUGINS,
-} from '../../googlesitekit/datastore/user/constants';
-import { getItem, setItem } from '../../googlesitekit/api/cache';
-import SpinnerButton from '../SpinnerButton';
-import ErrorNotice from '../ErrorNotice';
-import BannerNotification from './BannerNotification';
+} from '../../../googlesitekit/datastore/user/constants';
+import { getItem, setItem } from '../../../googlesitekit/api/cache';
+import SpinnerButton from '../../SpinnerButton';
+import ErrorNotice from '../../ErrorNotice';
+import BannerNotification from '../BannerNotification';
 
 const { useSelect, useDispatch } = Data;
